requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
gunicorn==23.0.0
azure-storage-blob==12.23.1
python-multipart==0.0.18